requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=0_gnu
blas=1.0=mkl
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py38h32f6830_0
cycler=0.10.0=py_2
freetype=2.10.2=he06d7ca_0
intel-openmp=2019.4=243
jpeg=9d=h516909a_0
kiwisolver=1.2.0=py38hbf85e49_0
lcms2=2.11=hbd6801e_0
ld_impl_linux-64=2.34=hc38a660_9
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.2.0=h24d8f2e_2
libgfortran-ng=7.5.0=hdf63c60_10
libgomp=9.2.0=h24d8f2e_2
libpng=1.6.37=hed695b0_1
libstdcxx-ng=9.2.0=hdf63c60_2
libtiff=4.1.0=hc7e4089_6
libwebp-base=1.1.0=h516909a_3
lz4-c=1.9.2=he1b5a44_1
matplotlib=3.3.0=1
matplotlib-base=3.3.0=py38h91b0d89_1
mkl=2019.4=243
mkl-service=2.3.0=py38h516909a_0
mkl_fft=1.1.0=py38hc1659b7_1
mkl_random=1.1.0=py38hb3f55d8_0
ncurses=6.2=he1b5a44_1
ninja=1.10.0=hc9558a2_0
numpy=1.18.5=py38ha1c710e_0
numpy-base=1.18.5=py38hde5b4d6_0
olefile=0.46=py_0
openssl=1.1.1g=h516909a_0
pandas=1.0.5=py38hcb8c335_0
patsy=0.5.1=py_0
pillow=7.2.0=py38h9776b28_1
pip=20.1.1=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.8.5=h425cb1d_1_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytz=2020.1=pyh9f0ad1d_0
readline=8.0=he28a2e2_2
scipy=1.5.0=py38h0b6359f_0
seaborn=0.10.1=1
seaborn-base=0.10.1=py_1
setuptools=49.2.0=py38h32f6830_0
six=1.15.0=pyh9f0ad1d_0
sqlite=3.32.3=hcee41ef_1
statsmodels=0.11.1=py38h1e0a361_2
tk=8.6.10=hed695b0_0
toml=0.10.1=pyh9f0ad1d_0
tornado=6.0.4=py38h1e0a361_1
tqdm=4.48.0=pyh9f0ad1d_0
wheel=0.34.2=py_1
xz=5.2.5=h516909a_1
zlib=1.2.11=h516909a_1006
zstd=1.4.5=h6597ccf_2
